+++ src/components/search.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { useLocalStorage } from './helpers';
+import { useLocalStorage } from '../hooks';
 
 export const Search = (props: {
   onSearch: (value: string) => Promise<void>;
