+++ src/hooks/index.tsx
@@ -1,3 +1,2 @@
 export { useLocalStorage } from './useLocalStorage';
 export { usePagination } from './usePagination';
-export { downloadSelectedItems } from './download-selected-items';
