+++ src/hooks/usePagination.tsx
@@ -1,6 +1,6 @@
 import { useSearchParams } from 'react-router';
 
-import { OFFSET_KEY, LIMIT_KEY, LIMIT_NUMBER } from './consts';
+import { OFFSET_KEY, LIMIT_KEY, LIMIT_NUMBER } from '../../constants';
 
 export function usePagination(): [
   curPage: string,
